Add two new methods: 1. _get_rowwise_agent_obstacle_diffs This method is an expansion of _get_rowwise_diffs. The main difference is that this method assumes the moving obstacle has a different size compared to the agent (for example, an AV interacting with a cyclist).
2. check_collision_moving_bicycle
This method checks for a collision between our AV and a moving bicycle. It is an expansion of the check_collision_moving_cars method, where collision checking is performed by calculating the minimum distance between two vehicles using 2 * car_dimensions.radius. However, if the vehicles are different, the minimum distance calculation needs to use the sum of the agent’s radius and the obstacle’s radius.